+++ lib/core/errors/error_enums/error_enums.dart
@@ -1,5 +1,6 @@
+export 'archive_errors.dart';
 export 'auth_errors.dart';
-export 'main_db_errors.dart';
 export 'file_system_errors.dart';
+export 'main_db_errors.dart';
 export 'network_errors.dart';
 export 'validation_errors.dart';