+++ lib/core/errors/errors.dart
@@ -1,2 +1,3 @@
 export 'app_error.dart';
 export 'error_enums/error_enums.dart';
+export 'extensions/extensions.dart';