+++ lib/core/errors/extensions/extensions.dart
@@ -1,0 +1,1 @@
+export 'file_error_to_app_error_extension.dart';