+++ lib/db_core/new/services/archive_service/models/archive_params.dart
@@ -1,0 +1,15 @@
+import 'dart:isolate';
+
+class ArchiveParams {
+  final String storePath;
+  final String outputPath;
+  final String? password;
+  final SendPort sendPort;
+
+  ArchiveParams({
+    required this.storePath,
+    required this.outputPath,
+    this.password,
+    required this.sendPort,
+  });
+}