+++ lib/db_core/new/services/archive_service/models/archive_progress_message.dart
@@ -1,0 +1,7 @@
+class ArchiveProgressMessage {
+  final int current;
+  final int total;
+  final String fileName;
+
+  ArchiveProgressMessage(this.current, this.total, this.fileName);
+}